requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
dj-database-url==3.0.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-environ==0.12.0
docopt==0.6.2
idna==3.10
lxml==5.4.0
pipreqs==0.4.13
psycopg2-binary==2.9.10
PyMuPDF==1.26.0
python-docx==1.1.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
yarg==0.1.10
